Make several modifications to divorce/decouple the Game model from any dependency on or knowledge of the canvas it will be drawn to or its dimensions.  These modifications were necessary to allow for the creation or restoration of a Game from the Game Fragment to be passed down through the various models and views where it is used.  Otherwise, attempting to create the game within the Game Fragment and hand it off to the Game SurfaceView presented a chicken/egg problem where the canvas size needed to be known by the Game model before being handed off to the SurfaceView constructor, but the SurfaceView needed the game during construction:
Modified the Fragment Game Layout, replacing the statically specified Game SurfaceView with a Frame.

Modified the Game Fragment to create a Game, create a Game SurfaceView, passing the game into the constructor, and then add the Game SurfaceView to the frame.  This also allowed for a custom constructor in the Game SurfaceView to accept a Game instead of resorting to Interfaces and Exceptions to do the same work.

Modified the Game SurfaceView to accept a Game in its constructor and pass it along to the Game Thread it uses.

Modified Game Thread to accept a Game in its constructor and pass it along to the Game Arena with the canvas to which its elements will be drawn.

Created a new Segment model to specify the Centipede/Segments in the Game based on percentages of whatever canvas to which they will be drawn.  This model is yet incomplete with respect to certain convenience functions the previous version offered, but the few steps back allow for leaps forward.

Modified Game to establish the presence of one Above Ground Centipede/Segment, and one Below Ground one, based on percentages of one/unit.  It is assumed/understood that the canvas will be 7:11 dimension ratio.

Modified Layer Above and Layer Below to draw their respective elements of the Game to the canvas using the new Segment's methods for resolving actual sizes from the percentages based on the actual canvas size.