Added Exit model that maintains directions on the X and Y axes, nominally as 1, 0, or -1, for positive, stationary, or negative movement respectively.  Exit requires each direction upon creation, and exposes getters to retrieve them.  Also provided are static convenience methods that return arrays of Exits in various directions intended to be provided to Turnstiles.
Added Turnstile model that maintains a current position percentages along the X and Y axes, a temporary radius percentage for drawing to Canvas during testing, and an ArrayList of Exits.  These are all required to be provided upon creation.  Getters for current position and radius on a Canvas are also provided, much as those provided by Segment.  Also provided is a method to get a random Exit from its Exits.

Modified Segment, removing updatePosition( double ) and adding several getters and setters for the various members modified there.  The update functionality was moved into Game in preparation for the direction changing logic keyed off of Turnstiles.

Modified Game, changing the concept of Velocity to Speed as Direction is now handled separately.  Added logic to add Turnstiles to the Game arena and temporary logic to draw them to the Canvas to ease testing.