Modified the public methods of Exit to package private.
Modified Game.  Adjusted the radius of Turnstiles to half their original size, added solid colors with which they are drawn to the Canvas, and discovered and corrected an issue with not all desired spots having one.  Added additional logic to animateSegmentThroughTurnstiles, calling it after animateSegmentThroughHoles.  This logic is not yet working as anticipated.

Modified Segment to track the Exit it took from the last Turnstile it passed through, providing the necessary getters and setters.

Modified Turnstile with getters for its current position percentage on the X and Y axes and what exits it has.  That latter is temporary to aid coloring it correctly when drawn to Canvas during testing.