Modified Game Fragment to publish Score and Clock values from the Game to the scoreboard when told to do so by the Game Thread.
Modified Game to track Score and Time Remaining, publishing updates to the score when Segments are successfully attacked.  Adjusted attackCentipedes not to kill segments that are touched when the game is paused.

Modified Game Thread to accept a reference to the Game Fragment, calling its onGameStatsChanged after running the game loop.

Modified Game Surface View to accept and pass a Game Fragment to its Game Thread for the aforementioned scoring.