Modified Game Activity to treat Back/Up Navigation from Game Fragment specially, deferring to Game Fragment to specify via boolean whether it should continue.
Modified Game Fragment to track whether Back/Up was tapped or not, preventing Back/Up Navigation on the first tap, requiring a second tap.  Additional functionality to display a dialog after the first tap is forthcoming.  Moved logic to toggle the Play/Pause button into a method so it could be called when the Game Fragment is stopped.  Modified onStop to pause the game so it does not jump wildly from a large time gap upon return.

Modified Game to expose a Pause method.

Made Game Fragment local to Game SurfaceView constructor as it was not used elsewhere.

Corrected README.md typo.