Modified Game Activity to forward activity results to the superclass for delivery to fragments that will consume them.
Modified Play Game Fragment heavily.  Implemented functionality to start a Google Sign-in intent on the Google Sign-in button, forwarding the signed-in account (or null) to the Game Fragment.

Modified Game Fragment to check for a Google Sign-in account and display either a Guest avatar and display name, or the signed in player's avatar and display name.

Added an icon/avatar for guest players and an overlay to enclose the guest or player avatar within a circle for aesthetics.

Modified Fragment Game layout to display the guest avatar with overlay and switched the padding to margins for proper display.

Modified the Game Fragment in Navigation Game to accept a Google Sign-in account.

Added Internet and Network Status permissions to the manifest to support access to Google Sign-in and Firebase services, and to support image retrieval with Glide in Game Fragment.

Added the Google Services JSON file provided by the Firebase application created on the console.

Modified the Gradle build files to include the dependencies necessary for the above modifications.