Added vector drawable assets for the game's earth layer, grass layer, and above and below ground centipede segments, differentiating heads from tails with eyes.
Modified Game to preload the drawable assets as bitmaps scaled to the appropriate size, drawing these to canvas.  This removed the need for initializing colors and paints, and removed the need to process many calculations in drawing portion of the loop.  As a result, frames render much more quickly, and the game can achieve a higher maximum speed.