Modified Direction, Position, and Moving Game Element to use floats again instead of integers.
Modified Game to capture canvas dimensions on initialization and draw screen elements based on percentages of the dimensions again.  Reduced the maximum speed to accommodate calculation of position in game loop.

Modified Game Thread to pass the canvas to Game on initialization instead of its width.  Game gets the dimensions from it internally.